// apps/rs-marketplace/Marketplace/RS-CDS.-Oficial/types.ts
export type CDType = 'PROPRIO' | 'FRANQUIA' | 'HIBRIDO';

export interface CDProfile {
  id: string;
  name: string;
  type: CDType;
  managerName: string;
  avatarUrl?: string; // Novo campo para foto de perfil
  region: string;
  walletBalance: number;
  activeCustomers: number;
  monthlyCycles: number;
}
// ...
export interface OrderDetail {
  productId: string;
  productName: string;
  quantity: number;
  unitPrice: number;
  points: number; // Pontos gerados por este item
}

export interface Order {
  id: string;
  consultantName: string;
  consultantPin: string;
  sponsorName?: string; // Nome do Patrocinador
  sponsorId?: string;   // ID do Patrocinador
  buyerCpf?: string;
  buyerEmail?: string;
  buyerPhone?: string;
  shippingAddress?: string;
  date: string;
  time?: string;        // Horário do pedido/retirada
  total: number;
  totalPoints: number;  // Total de VP do pedido
  status: 'PENDENTE' | 'SEPARACAO' | 'AGUARDANDO_RETIRADA' | 'EM_TRANSPORTE' | 'CONCLUIDO';
  type: 'RETIRADA' | 'ENTREGA';
  items: number;
  trackingCode?: string;
  vehiclePlate?: string; // Placa do veículo de retirada/entrega
  productsDetail?: OrderDetail[]; // Lista detalhada de itens
}
// ...
export type ViewState = 'DASHBOARD' | 'PEDIDOS' | 'ESTOQUE' | 'FINANCEIRO' | 'IA_ADVISOR' | 'CONFIGURACOES' | 'HISTORICO' | 'HEADQUARTERS';